Fixed visual studio compilation problems and removed a warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1777 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/grand_prix_select.cpp
@@ -106,7 +106,7 @@ void GrandPrixSelect::update(float dt)
         !( SELECTED_TOKEN < WTOK_FIRSTPRIX ))
     {
         const int CUP_NUM = SELECTED_TOKEN - WTOK_FIRSTPRIX;
-        const int NUM_TRACKS = m_all_cups[CUP_NUM].getTrackCount();
+        const int NUM_TRACKS = (int)m_all_cups[CUP_NUM].getTrackCount();
 
         const CupData &cup = m_all_cups[CUP_NUM];
         widget_manager->setWgtText(WTOK_DESCRIPTION, cup.getDescription());
